mac/requirements.txt
Django==1.6.2
Pillow==2.0.0
South==0.8.4

django-cms==3.0
django-classy-tags==0.4
django-diplomat==0.2.0
django-mandrill==0.1.0
django-mptt==0.6
django-phonenumber-field==0.5
django-reversion==1.7
django-sekizai==0.7
djangocms-admin-style==0.1.2
djangorestframework==2.3.13
django-filer==0.9.5

cmsplugin-filer==0.9.5
html5lib==1.0b1
psycopg2==2.5
six==1.3.0
